agents/研发博士/log-analysis/requirements.txt
# 日志智能分析 - 依赖包

# 核心依赖
pyyaml>=6.0
requests>=2.28.0

# 可选: 增强功能
# openpyxl>=3.0.0    # Excel支持
# pandas>=1.5.0     # 数据分析
